[celery] refactor Celery test cases
